% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirFireModeling.R
\docType{package}
\name{AirFireModeling}
\alias{AirFireModeling}
\alias{AirFireModelingEnv}
\title{Utilities for working with USFS AirFire BlueSky model output}
\format{
An object of class \code{environment} of length 1.
}
\usage{
AirFireModelingEnv
}
\description{
A suite of utility functions providing functionality commonly
needed for working with BlueSky model output in R.
}
\keyword{datasets}
