% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluesky_download.R
\name{bluesky_download}
\alias{bluesky_download}
\title{Download BlueSky model data from AirFire}
\usage{
bluesky_download(
  modelName = "PNW-4km",
  modelRun = NULL,
  baseUrl = "https://haze.airfire.org/bluesky-daily/output/standard",
  modelMode = "forecast",
  verbose = TRUE
)
}
\arguments{
\item{modelName}{Model identifier.}

\item{modelRun}{Model initialization datestamp as "YYYYMMDDHH".}

\item{baseUrl}{Base URL for BlueSky output.}

\item{modelMode}{Subdirectory path containing BlueSky output, i.e. 'forcast'.}

\item{verbose}{If \code{FALSE}, suppress status messages (if any), and the
progress bar.}
}
\value{
File path of downloaded data.
}
\description{
Downloads a copy of the specified BlueSky model run to the
package data directory. This file can then be converted into a common format.

#' On 2019-10-11, available model identifiers include the following:
\itemize{
  \item{AK-12km}
  \item{CANSAC-1.33km}
  \item{CANSAC-4km}
  \item{DRI1.33km-CMAQ}
  \item{DRI4km-CMAQ}
  \item{GFS-0.15deg-CanadaUSA-p25deg-68N}
  \item{GFS-0.15deg}
  \item{NAM-3km}
  \item{NAM84-0.15deg}
  \item{PNW-1.33km}
  \item{PNW-4km}
  \item{PNW1.33km-CMAQ}
  \item{PNW4km-CMAQ}
}

Users will typically call bluesky_load() which in turn calls this function.

BlueSky output files are found in directories with the following
structure: \code{<baseUrl>/<modelName>/<modelRun>/<modelMode>/data/...}

\preformatted{
<baseUrl>/NAM84-0.08deg/2016050600/carryover/data/...
<baseUrl>/NAM84-0.08deg/2016050600/combined/data/...
<baseUrl>/NAM84-0.08deg/2016050600/forecast/data/...
}
}
\examples{
\donttest{
library(AirFireModeling)
setModelDataDir('~/Data/BlueSky')

filePath <- bluesky_download(modelName = "PNW-4km", modelRun = 2020091300)
bluesky_toCommonFormat(filePath)
bluesky_downloaded()
}
}
\seealso{
\link{setModelDataDir}
}
