% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirFireModeling.R
\name{bluesky_downloaded}
\alias{bluesky_downloaded}
\title{List Downloaded Models}
\usage{
bluesky_downloaded(
  path = getModelDataDir(),
  pattern = ".nc",
  full = FALSE,
  ...
)
}
\arguments{
\item{path}{The path in which models are located.}

\item{pattern}{A regex pattern to use for filtering model files}

\item{full}{Logical. Show the full path of the model (used for local loading).}

\item{...}{Additional arguments to be passed to \code{list.files()}.}
}
\value{
A list of downloaded models
}
\description{
List Downloaded Models
}
