% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluesky_findModels.R
\name{bluesky_findModels}
\alias{bluesky_findModels}
\title{Find models covering a location}
\usage{
bluesky_findModels(longitude = NULL, latitude = NULL)
}
\arguments{
\item{longitude}{the target longitude}

\item{latitude}{the target latitude}
}
\value{
Vector of model names.
}
\description{
Searches \code{bluesky_modelInfo} for all models whose domain
includes the incoming location.
}
\examples{
library(AirFireModeling)
setModelDataDir('~/Data/BlueSky')

# Models covering Portland, Oregon
bluesky_findModels(
  longitude = -122.68,
  latitude = 45.52
)

}
