% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluesky_latestModelRun.R
\name{bluesky_latestModelRun}
\alias{bluesky_latestModelRun}
\title{Find latest BlueSky model run from AirFire}
\usage{
bluesky_latestModelRun(
  modelName = NULL,
  count = 1,
  baseUrl = "https://haze.airfire.org/bluesky-daily/output/standard",
  verbose = TRUE
)
}
\arguments{
\item{modelName}{Model identifier.}

\item{count}{Number of recent \code{modelRun} strings to return.}

\item{baseUrl}{Base URL for BlueSky output.}

\item{verbose}{If \code{FALSE}, suppress status messages (if any), and the
progress bar.}
}
\value{
Timestamp of the latest model run.
}
\description{
Scans the directory of BluSky model output and returns the
most recent \code{modelRun} string(s). If \code{count > 1}, the most recent
\code{count} will be returned in low-hi order.

#' On 2019-10-11, available model identifiers include the following:
\itemize{
  \item{AK-12km}
  \item{CANSAC-1.33km}
  \item{CANSAC-4km}
  \item{DRI1.33km-CMAQ}
  \item{DRI4km-CMAQ}
  \item{GFS-0.15deg-CanadaUSA-p25deg-68N}
  \item{GFS-0.15deg}
  \item{NAM-3km}
  \item{NAM84-0.15deg}
  \item{PNW-1.33km}
  \item{PNW-4km}
  \item{PNW1.33km-CMAQ}
  \item{PNW4km-CMAQ}
}

Users wanting the latest data will typically call this function first,
followed by bluesky_load().

BlueSky output files are found in directories with the following
structure: \code{<baseUrl>/<modelName>/<modelRun>/<modelMode>/data/...}

\preformatted{
<baseUrl>/NAM84-0.08deg/2016050600/carryover/data/...
<baseUrl>/NAM84-0.08deg/2016050600/combined/data/...
<baseUrl>/NAM84-0.08deg/2016050600/forecast/data/...
}
}
\examples{
\donttest{
library(AirFireModeling)
setModelDataDir('~/Data/BlueSky')

modelRun <- bluesky_latestModelRun(modelName = "PNW-4km")
filePath <- bluesky_download(modelName = "PNW-4km", modelRun = modelRun)
bluesky_toCommonFormat(filePath)
bluesky_downloaded()
}
}
