% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluesky_load.R
\name{bluesky_load}
\alias{bluesky_load}
\title{Load BlueSky NetCDF file and return contents as a bs_grid.}
\usage{
bluesky_load(dailyOutputDir = "standard", model = "PNW-1.33km",
  modelRun = NULL, subDir = "combined", parameter = "pm25",
  download = TRUE, cleanup = TRUE, filePath = NULL,
  baseUrl = "https://haze.airfire.org/bluesky-daily/output",
  quiet = FALSE)
}
\arguments{
\item{dailyOutputDir}{BlueSky web directory.}

\item{model}{Model identifier.}

\item{modelRun}{Date code as "YYYYMMDDHH" (integer or character).}

\item{subDir}{Subdirectory path containing netcdf data.}

\item{parameter}{Parameter name.}

\item{download}{Logical specifying whether to download and convert data if it
is not found locally.}

\item{cleanup}{Logical specifying whether to remove the original files.}

\item{filePath}{Absolute path of the local bluesky .nc file 
\emph{(not "_V2.nc")} for special occasions.}

\item{baseUrl}{Base URL for BlueSky output.}

\item{quiet}{If \code{TRUE}, suppress status messages (if any), and the
progress bar.}
}
\value{
List of class \code{bs_grid} with the following elements:
\itemize{
  \item{lon -- 1D vector of longitudes (X)}
  \item{lat -- 1D vector of latitudes (Y)}
  \item{time -- 1D vector of POSIXct times (T)}
  \item{pm25 -- 3D array of PM2.5 values (X-Y-T)}
  \item{model -- character name of model used to generate the data}
  \item{modelRun -- character date + hour when the model was run}
  \item{deltaLon -- longitude axis grid cell spacing}
  \item{deltaLat -- latitude axis grid cell spacing}
}

BlueSky output files are found in directories with the following
structure:

\preformatted{
.../standard/NAM84-0.08deg/2016050600/carryover/data/...
.../standard/NAM84-0.08deg/2016050600/combined/data/...
.../standard/NAM84-0.08deg/2016050600/forecast/data/...
}
}
\description{
Reads in a "_V2" BlueSky output NetCDF file and ingests the 
data, returning it in a list of class \code{bs_grid}.
}
\details{
Data are searched for in the package internal \code{dataDir} which 
can be set with \code{setModelDataDir()}.

#' On 2019-10-11, vailable model identifiers include the following:
\itemize{
  \item{AK-12km}
  \item{CANSAC-1.33km}
  \item{CANSAC-4km}
  \item{DRI1.33km-CMAQ}
  \item{DRI4km-CMAQ}
  \item{GFS-0.15deg-CanadaUSA-p25deg-68N}
  \item{GFS-0.15deg}
  \item{NAM-3km}
  \item{NAM84-0.15deg}
  \item{PNW-1.33km}
  \item{PNW-4km}
  \item{PNW1.33km-CMAQ}
  \item{PNW4km-CMAQ}
}
}
\examples{
\dontrun{
setModelDataDir('~/Data/Bluesky')
bs_grid <- bluesky_load(model = "PNW-1.33km", modelRun = 2019090900)
}
}
\seealso{
\link{bluesky_download}

\link{bluesky_toV2}

\link{setModelDataDir}
}
