% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluesky_locationPlot.R
\name{bluesky_locationPlot}
\alias{bluesky_locationPlot}
\title{Downlaod and Plot BlueSky location models}
\usage{
bluesky_locationPlot(
  modelRun,
  latitude = NULL,
  longitude = NULL,
  monitorID = NULL,
  models = c("PNW-1.33km", "PNW-4km"),
  ...
)
}
\arguments{
\item{modelRun}{Date code as "YYYYMMDDHH".}

\item{latitude}{A target latitude, optional if \code{monitorID} provided}

\item{longitude}{A target longitude, optional if \code{monitorID} provided}

\item{monitorID}{(Optional) a monitorID to plot along same model run period.}

\item{models}{A list of models to plot}

\item{...}{additional parameters passed into \code{bluesky_load()}}
}
\value{
A ggplot object
}
\description{
Downlaod and Plot BlueSky location models
}
\examples{
\dontrun{
monitorID = '410650007_01'
bluesky_locationPlot(
  modelRun = 20200301, 
  monitorID = monitorID, 
  models = c('PNW-1.33km', 'PNW-4km', 'CANSAC-4km')
)
}
}
