% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirFireModeling.R
\docType{data}
\name{bluesky_modelInfo}
\alias{bluesky_modelInfo}
\title{BlueSky Model Information}
\format{A data frame}
\usage{
bluesky_modelInfo
}
\description{
Various information regarding BlueSky models.
}
\details{
blues_modelInfo
}
\keyword{datasets}
