% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluesky_spaghetti.R
\name{bluesky_spaghetti}
\alias{bluesky_spaghetti}
\title{Spaghetti Plot}
\usage{
bluesky_spaghetti(raster, longitude = NULL, latitude = NULL, radius = 5000)
}
\arguments{
\item{raster}{A Raster* object}

\item{longitude}{A target latitude}

\item{latitude}{A target longitude}

\item{radius}{A radial distance about the target location}
}
\value{
a gg object
}
\description{
Plot a spaghetti plot of all model adjacent cells to a target
location
}
