% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluesky_toCommonFormat.R
\name{bluesky_toCommonFormat}
\alias{bluesky_toCommonFormat}
\title{Convert BlueSky NetCDF output to modernized NetCDF}
\usage{
bluesky_toCommonFormat(filePath = NULL, clean = TRUE)
}
\arguments{
\item{filePath}{Absolute path of file to be converted.}

\item{clean}{Logical specifying whether to remove the original netcdf file.}
}
\value{
Absolute path of the converted NetCDF file.
}
\description{
Converts BlueSky model output from its original format to a more
modernized NetCDF format with dimension axes for longitude, latitude,
elevation and time. With default settings, output files renamed to
~base~_v2.nc.
}
\note{
Users will typically call \code{bluesky_load()} which in turn calls
this function.
}
\examples{
\dontrun{
library(AirFireModeling)
setModelDataDir('~/Data/BlueSky')

filePath <- bluesky_download(model = "PNW-4km", modelRun = 2019100900)
bluesky_toCommonFormat(filePath)
bluesky_downloaded()
}
}
