% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluesky_toV2.R
\name{bluesky_toV2}
\alias{bluesky_toV2}
\title{Convert BlueSky NetCDF output to modernized NetCDF}
\usage{
bluesky_toV2(filePath = NULL, cleanup = TRUE)
}
\arguments{
\item{filePath}{Absolute path of file to be converted.}

\item{cleanup}{Logical specifying whether to remove the original netcdf file.}
}
\value{
Absolute path of the converted NetCDF file.
}
\description{
Converts BlueSky model output from its original format to a more 
modernized NetCDF format with dimension axes for longitude, latitude, 
elevation and time. With default settings, output files renamed to 
~base~_V2.nc.
}
\note{
Users will typically call \code{bluesky_load()} which in turn calls 
this function.
}
\examples{
\dontrun{
setModelDataDir('~/Data/Bluesky')
filePath <- bluesky_download(model = "PNW-4km", modelRun = 2019100900)
bluesky_toV2(filePath)
}
}
