% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirFireModeling.R
\name{bluesky_whichModel}
\alias{bluesky_whichModel}
\title{Find coordinates appropriate model}
\usage{
bluesky_whichModel(longitude, latitude)
}
\arguments{
\item{longitude}{the target longitude}

\item{latitude}{the target latitude}
}
\value{
vectors of model(s)
}
\description{
Determine what model to use based on the target coordinates
provided.
}
