% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmaq_load.R
\name{cmaq_load}
\alias{cmaq_load}
\title{Load data for a single CMAQ model}
\usage{
cmaq_load(
  localPath = NULL,
  level = 1,
  xlim = NULL,
  ylim = NULL,
  clean = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{localPath}{Absolute path to a NetCDF file not found in `modelDataDir`.}

\item{level}{Elevation level.}

\item{xlim}{A vector of coordinate longitude bounds.}

\item{ylim}{A vector of coordinate latitude bounds.}

\item{clean}{Logical specifying removal of original model data after conversion
to "v2" format.}

\item{verbose}{Logical to display messages.}
}
\value{
A \pkg{raster} package \emph{RasterBrick} object.
}
\description{
This function encapsulates the process of downloading, and formatting
necessary for CMAQ model outputs. Additionally, it checks already existing
models and loads those if available.

Loads CMAQ model output as a raster brick. The directory
previously set with \code{setModelDataDir()} is searched for "v2" formatted
model output. If data are not found locally, data are downloaded from
\code{baseUrl} and converted.

The returned \code{dataBrick} object is of class \code{raster::RasterBrick} and
can be manipulated with appropriate functions from the \pkg{raster} package
or any of the \code{raster_~()} functions provided by \pkg{AirFireModeling}.
}
\note{
The regurned \emph{RasterBrick} object will contain all grid cells
within the area defined by \code{xlim} and \code{ylim}.

If \code{localPath} is used, \code{clean = FALSE} is automatically set.
The package will never remove original model output that it has not downloaded.
}
