% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusionMatrix.R
\name{confusionMatrix}
\alias{confusionMatrix}
\title{Confusion Matrix}
\usage{
confusionMatrix(predicted, observed)
}
\arguments{
\item{predicted}{Preicted logical.}

\item{observed}{Observed logical.}
}
\value{
a list of contingency statistics
}
\description{
Confusion Matrix
}
