% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-raster.R
\name{createLayerTime}
\alias{createLayerTime}
\title{Create raster layer times}
\usage{
createLayerTime(layerName = "")
}
\arguments{
\item{layerName}{Name of the \code{rasterLayer} object}
}
\value{
\code{POSIXct} time.
}
\description{
Create raster layer times
}
\keyword{internal}
