% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_raster.R
\name{createLayerTimeStamp}
\alias{createLayerTimeStamp}
\title{Create raster layer timestamp}
\usage{
createLayerTimeStamp(layerName = "")
}
\arguments{
\item{layerName}{Name of the \code{rasterLayer} object}

\item{timezone}{Olson timezone in which times will be displayed.}

\item{prefix}{String prepended to the time}
}
\value{
UTC timestamp as YYYYmmddHH.
}
\description{
Create raster layer timestamp
}
\keyword{internal}
