% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-raster.R
\name{createLayerTimeString}
\alias{createLayerTimeString}
\title{Create raster layer time string}
\usage{
createLayerTimeString(layerName = "", timezone = "UTC", prefix = "")
}
\arguments{
\item{layerName}{Name of the \code{rasterLayer} object}

\item{timezone}{Olson timezone in which times will be displayed.}

\item{prefix}{String prepended to the time}
}
\value{
Time string formatted for \code{timezone}.
}
\description{
Create raster layer time string
}
