% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirFireModeling.R
\name{.bluesky_load}
\alias{.bluesky_load}
\title{Internal Model Load Function}
\usage{
.bluesky_load(model, run, xlim, ylim, local, dirURL, type, clean, verbose)
}
\arguments{
\item{model}{A model. i.e 'PNW-1.33km' or 'CANSAC-4km'.}

\item{run}{The model run date. YYYYmmddHH format required.}

\item{xlim}{A vector of coordinate longitude bounds.}

\item{local}{A path to a downloaded NetCDF model.}

\item{clean}{Option to clean, or delete, the non-formatted model.}

\item{verbose}{Logical to display messages.}

\item{ylime}{A vector of coordinate latitude bounds.}

\item{dirUrl}{the database URL.}
}
\value{
A RasterBrick
}
\description{
This function encapsulates the process of downloading, and formatting
necessary for BlueSky model outputs. Additionally, it checks already existing
models and loads those if avaliable.
}
\keyword{internal}
