% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-grid.R
\name{grid_boundingBox}
\alias{grid_boundingBox}
\title{Find bs_grid bounding box}
\usage{
grid_boundingBox(bs_grid)
}
\arguments{
\item{bs_grid}{a bs_grid object}
}
\value{
a bounding box matrix
}
\description{
Quickly find the bounding box of a \code{bs_grid} object.
}
