% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_correlationMap.R
\name{grid_correlationMap}
\alias{grid_correlationMap}
\title{Correlation Map}
\usage{
grid_correlationMap(bs_grid = NULL, ws_monitor = NULL,
  monitorID = NULL, parameter = "pm25", xlim = NULL, ylim = NULL,
  tlim = NULL, breaks = c(-1, -0.7, -0.5, -0.3, -0.1, 0.1, 0.3, 0.5,
  0.7, 1), colors = rev(RColorBrewer::brewer.pal(9, "RdBu")),
  zoom = 1.5, plotter = "base")
}
\arguments{
\item{bs_grid}{bs_grid data list}

\item{ws_monitor}{monitor data list}

\item{monitorID}{monitor ID for correlation to model data}

\item{parameter}{Data parameter to plot.}

\item{xlim}{longitude range to be plotted}

\item{ylim}{latitude range to be plotted}

\item{tlim}{Time range tover which to calculation the correlation interpreted
as UTC.}

\item{breaks}{used by \code{image()} function to partition correlation values 
into different colors}

\item{colors}{vector of colors to use}

\item{zoom}{range extension as a fraction of the coordinate range}

\item{plotter}{either \code{"base"} or \code{"ggplot"}.}
}
\description{
Plots a map of the correlation of bluesky model data with a 
monitoring timeseries.

When \code{colors=NULL}, a color palette is given by \code{colorRampPalette}.
Note: when using `plotter = ggplot`, zoom, colors and breaks are invalidated.
}
\examples{
\dontrun{
library(PWFSLSmoke)
library(AirFireModeling)

setModelDataDir("~/Data/Bluesky")

ws_monitor <- monitor_load(20191014, 20191016) \%>\%
  monitor_subset(monitorIDs = "lon_.120.591_lat_38.714_arb2.1008")
bs_grid <- bluesky_load(model = "CANSAC-1.33km", modelRun = 20191014)
grid_correlationMap(bs_grid, ws_monitor)
}
}
\seealso{
\link{grid_map}
}
\keyword{bs_grid,}
\keyword{ws_monitor}
