% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_distance.R
\name{grid_distance}
\alias{grid_distance}
\title{Create distance-from-targetgGrid}
\usage{
grid_distance(bs_grid = NULL, longitude, latitude)
}
\arguments{
\item{bs_grid}{\emph{bs_grid} object.}

\item{longitude}{Longitude of the point from which distances are calculated
in decimal degrees E.}

\item{latitude}{Latitude of the point from which distances are calculated
in decimal degrees N.}
}
\value{
2d matrix of distances.
}
\description{
This function returns 2D matrix (aka 'grid'). Each cell in the 
grid contains the distance in meters between the center of that cell and the 
target location. This distance grid can be used to create a mask identifying 
grid cells within a certain radius of the target location.
}
\examples{
\donttest{
setModelDataDir("~/Data/Bluesky")
bs <- bluesky_load(model = "PNW-4km", modelRun = 2019100900)
xlim <- c(-124, -122)
ylim <- c(46, 48)
bs <- grid_subset(bs, xlim=xlim, ylim=ylim)
distanceMatrix <- grid_distance(bs, -123.801, 47.704)
image(bs$longitude, bs$latitude, distanceMatrix, xlab = "", ylab = "")
}
}
