% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_ggmap.R
\name{grid_ggmap}
\alias{grid_ggmap}
\title{Map gridded model data with ggplot2}
\usage{
grid_ggmap(bs_grid, parameter = "pm25",
  xlim = range(bs_grid$longitude), ylim = range(bs_grid$latitude),
  tMask = rep(TRUE, times = length(bs_grid$time)), slice = mean)
}
\arguments{
\item{bs_grid}{\emph{bs_grid} object}

\item{parameter}{Data parameter to plot.}

\item{xlim}{Longitude range to be plotted.}

\item{ylim}{Latitude range to be plotted.}

\item{tMask}{Logical vector of the same length as the time axis or a set
of indices used to select specific times.}

\item{slice}{Either a time index or a function used to collapse the time axis.}
}
\description{
Use the \pkg{ggplot2} package to plot a map of bluesky data in a
given area.
}
\examples{
\dontrun{
setModelDataDir('~/Data/Bluesky')
bs_grid <- bluesky_load(model = "PNW-1.33km", modelRun = 2019100900, subDir = 'forecast')
grid_ggmap(bs_grid, xlim = xlim, ylim = ylim, tMask = 1:40, slice = max)
}
}
