% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-grid.R
\name{grid_isGrid}
\alias{grid_isGrid}
\title{Test for correct structure in a \emph{bs_grid} object}
\usage{
grid_isGrid(bs_grid = NULL)
}
\arguments{
\item{bs_grid}{\emph{bs_grid} object}
}
\value{
\code{TRUE} if \code{bs_grid} has the correct structure,
\code{FALSE} otherwise.
}
\description{
The \code{bs_grid} is checked for the 'bs_grid' class name
and presence of the following elements:
\itemize{
  \item{longitude}
  \item{latitude}
  \item{elevation}
  \item{time}
  \item{data}
  \item{model}
  \item{modelRun}
  \item{deltaLon}
  \item{deltaLat}
}
}
