% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_map.R
\name{grid_map}
\alias{grid_map}
\title{Map gridded model data}
\usage{
grid_map(bs_grid, parameter = "pm25", xlim = range(bs_grid$longitude),
  ylim = range(bs_grid$latitude), tMask = rep(TRUE, times =
  length(bs_grid$time)), slice = max, style = "default",
  breaks = NULL, colors = NULL, addAxes = TRUE, main = NULL, ...)
}
\arguments{
\item{bs_grid}{\emph{bs_grid} object}

\item{parameter}{Data parameter to plot.}

\item{xlim}{Longitude range to be plotted.}

\item{ylim}{Latitude range to be plotted.}

\item{tMask}{Logical vector of the same length as the time axis or a set
of indices used to select specific times.}

\item{slice}{Either a time index or a function used to collapse the time axis.}

\item{style}{Custom styling, one of \code{'default'}.}

\item{breaks}{A set of numeric breakpoints for colors. This is passed on
to \code{graphics::image()} to partition values into different colors.}

\item{colors}{Vector of colors to use.}

\item{addAxes}{Logical specifying whether to add axes.}

\item{main}{Title for the plot.}

\item{...}{Aadditional arguments passed to maps::map().}
}
\description{
Use the \pkg{maps} package to plot a map of bluesky data in a 
given area.

When a named \code{style} is used, some graphical parameters will be 
overridden. Available styles include:

\itemize{
\item{\code{default}}{-- hourly values are individually colored by 24-hr AQI levels}
}

If \code{breaks} and \code{colors} are defined, they will override 
\code{style}.
}
\examples{
\donttest{
setModelDataDir('~/Data/Bluesky')
bs_grid <- bluesky_load(model = "PNW-1.33km", modelRun = 2019100900)
xlim <- c(-118, -114)
ylim <- c(45, 48)
grid_map(bs_grid, xlim = xlim, ylim = ylim, tMask = 1:40)
}
}
