% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_subset.R
\name{grid_subset}
\alias{grid_subset}
\title{Subset \code{bs_grid} to create a smaller grid}
\usage{
grid_subset(bs_grid, xlim = NULL, ylim = NULL, tlim = NULL)
}
\arguments{
\item{bs_grid}{\emph{bs_grid} data list.}

\item{xlim}{Optional vector of longitude lo and hi values.}

\item{ylim}{Optional vector of latitude lo and hi values.}

\item{tlim}{Optional vector of time lo and hi values in Ymd[HMS] format or
\code{POSIXct}.}
}
\value{
\code{bs_grid} object with reduced dimensions.
}
\description{
The incoming gridded data list is subset according to the 
parameters passed in.
}
\examples{
\donttest{
setModelDataDir('~/Data/Bluesky')
bs <- bluesky_load(model = "PNW-1.33km", modelRun = 2019100900)
xlim <- c(-118, -114)
ylim <- c(45, 48)
grid_map(bs, xlim = xlim, ylim = ylim, tMask = (1:24))
}
}
