% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_subsetByDistance.R
\name{grid_subsetByDistance}
\alias{grid_subsetByDistance}
\title{Subset a \emph{bs_grid} object by distance from a location}
\usage{
grid_subsetByDistance(bs_grid = NULL, longitude = NULL,
  latitude = NULL, radius = 20000, count = NULL)
}
\arguments{
\item{bs_grid}{\emph{bs_grid} object.}

\item{longitude}{Target longitude from which the distance will be calculated.}

\item{latitude}{Target latitude from which the distance will be calculated.}

\item{radius}{Radius in meters.}

\item{count}{Number of grid cells to return.}
}
\value{
\emph{ws_monitor} object where each unique monitor represents a
single grid cell from \code{bs_grid}.
}
\description{
This function subsets the \code{bs_grid} object and creates a
new object of type \code{ws_monitor} that is compatible with 
\pkg{PWFSLSmoke} \code{monitor_~} functions.

When \code{count} is used, a \code{ws_monitor} object is created containing 
\strong{up to} \code{count} monitors ordered by increasing distance from the 
target location. Only grid cells are found within \code{radius} will be 
returned even if this number is fewer than \code{count}.
}
\examples{
\donttest{
# Next three lines required for bs_grid to ws_monitor conversion
library(MazamaSpatialUtils)
setSpatialDataDir("~/Data/Spatial")
loadSpatialData("NaturalEarthAdm1")

# Now we can work as we normally would
setModelDataDir("~/Data/Bluesky")
bs <- bluesky_load(model = "PNW-1.33km", modelRun = 2019100900)
bs_monitors <- grid_subsetByDistance(bs, -121.017092, 46.901512, 
                                     radius = 10000)

# Now start using functions from PWFSLSMoke
bs_mean <- monitor_collapse(bs_monitors, monitorID = "mean", FUN = mean)
bs_median <- monitor_collapse(bs_monitors, monitorID = "median", FUN = median)
monitorPlot_timeseries(bs_monitors, style = "gnats", ylim = c(0, 100), cex = 1.5)
monitorPlot_timeseries(bs_median, type = "l", lwd = 3, col = "blue", add = TRUE)
monitorPlot_timeseries(bs_mean, type = "l", lwd = 3, col = "red", add = TRUE)
legend("topleft", col = c("red", "blue"), lwd = 1, legend = c("mean", "median"))
}
}
\seealso{
\link{grid_subsetByMask}
}
