% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_subsetByHUC.R
\name{grid_subsetByHUC}
\alias{grid_subsetByHUC}
\title{Subset a \emph{bs_grid} objectt by hydrologic unit code}
\usage{
grid_subsetByHUC(bs_grid, HUC = NULL)
}
\arguments{
\item{bs_grid}{\emph{bs_grid} object.}

\item{HUC}{USGS hydrologic unit code.}
}
\value{
\emph{ws_monitor} object where each unique monitor represents a
single grid cell from \code{bs_grid}.
}
\description{
This function subsets \code{bs_grid} and creates a new object
type \emph{ws_monitor}  that is compatible with \pkg{PWFSLSmoke} 
\code{monitor_~} functions.
}
\details{
Use of this function requires installation of the 
\code{WBDHU#.RData} files that can be downloaded with the 
\pkg{MazamaSpatialUtils} package. The appropriate data file matching the 
level of the \code{HUC} argument must be loaded as in the example.
}
\section{USGS Watershed Boundary Dataset}{

TODO:  Add description.
}

\examples{
\dontrun{
# Next three lines required for bs_grid to ws_monitor conversion
library(MazamaSpatialUtils)
setSpatialDataDir("~/Data/Spatial")
loadSpatialData("WBDHU10")

longitude <- -123.801
latitude <- 47.704
WA10 <- MazamaSpatialUtils::subsetHUC(WBDHU10, allStateCodes = "WA")
HUC <- MazamaSpatialUtils::getHUC(longitude, latitude, "WA10")

setModelDataDir("~/Data/Bluesky")
bs <- bluesky_load(model = "PNW-1.33km", modelRun = 2019100900)

QueetsRiver <- grid_subsetByHUC(bs, HUC)
PWFSLSmoke::monitor_staticmap(QueetsRiver, cex = .5, pch = 15)
}
}
\seealso{
\link{grid_subsetByMask}
}
