% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_subsetByMask.R
\name{grid_subsetByMask}
\alias{grid_subsetByMask}
\title{Subset a \emph{bs_grid} with with a gridMask}
\usage{
grid_subsetByMask(bs_grid = NULL, gridMask = NULL,
  countryCodes = c("CA", "US"))
}
\arguments{
\item{bs_grid}{\emph{bs_grid} object.}

\item{gridMask}{Logical 2D matrix of the same dimensions as the grid data.}

\item{countryCodes}{Vector of country codes covered by the grid.}
}
\value{
\emph{ws_monitor} object where each unique monitor represents a
single grid cell from \code{bs_grid}.
}
\description{
This function subsets the \code{bs_grid} object and 
restructures the data into a \emph{ws_monitor} that is compatible with 
\pkg{PWFSLSmoke} \code{monitor_~} functions.
}
\note{
The \code{countryCodes} parameter is provided because of the need
to perform spatial searches for during the conversion of grid cell locations into
\emph{ws_monitor} objects. Each location must have correct information for
\code{countryCode, stateCode, timezone} and these spatial searches can be
time consuming. Specifying the \code{countryCodes} covered by the grid
greatly speeds up these searches.
}
\examples{
\dontrun{
# Next three lines required for bs_grid to ws_monitor conversion
library(MazamaSpatialUtils)
setSpatialDataDir("~/Data/Spatial")
loadSpatialData("NaturalEarthAdm1")

# Now we can work as we normally would
setModelDataDir("~/Data/Bluesky")
bs <- bluesky_load(model = "PNW-1.33km", modelRun = 2019090900)
bs_monitors <- grid_subsetByDistance(bs, -121.017092, 46.901512, 
                                     radius = 10000)

distanceMatrix <- grid_distance(bs, -121.017092, 46.901512)
radius <- 20000
gridMask <- distanceMatrix <= radius
bs_monitors <- grid_subsetByMask(bs, gridMask)
PWFSLSmoke::monitor_staticmap(bs_monitors, cex = .5, pch = 15)
}
}
\seealso{
\link{grid_subsetByDistance}
}
