% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_timeseriesPlot.R
\name{grid_timeseriesPlot}
\alias{grid_timeseriesPlot}
\title{Timeseries Plot for a `bs_grid` Object}
\usage{
grid_timeseriesPlot(bs_grid, tlim = NULL, localTime = FALSE,
  shadedNight = FALSE, add = FALSE, gridPos = "",
  gridCol = "black", gridLwd = 1, gridLty = "solid", dayLwd = 0,
  hourLwd = 0, hourInterval = 6, title = NULL, ylim = NULL)
}
\arguments{
\item{bs_grid}{'gridded' BlueSky DataList}

\item{tlim}{optional vector with start and end times (integer or character representing YYYYMMDD[HH])}

\item{localTime}{A logical whether you are going to use GMT time or not}

\item{shadedNight}{A logical whether you want to add shaded rectangles for the local night times}

\item{add}{logical specifying whether to add points to an existing plot}

\item{gridPos}{position of grid lines either 'over', 'under' ('' for no grid lines)}

\item{gridCol}{grid line color}

\item{gridLwd}{grid line width}

\item{gridLty}{grid line type}

\item{dayLwd}{day marker line width}

\item{hourLwd}{hour marker line width}

\item{hourInterval}{interval for grid (max=12)}

\item{title}{The desired title of your plot. Setting as NULL prints out the default title.}

\item{ylim}{the y limits of the plot}
}
\description{
This funciton plot timeseries from an 'unraveled' BlueSky dataList.
}
\examples{
\dontrun{
  setModelDataDir('~/Data/Bluesky')
  bs_grid <- bluesky_load(model = "PNW-1.33km", modelRun = 2019100900)
  grid_timeseriesPlot(bs_grid, ylim = c(0,500))
}
}
\seealso{
\link{grid_subsetByDistance}
}
\keyword{bs_grid}
