% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_spaghettiPlot.R
\name{gridPlot_spaghetti}
\alias{gridPlot_spaghetti}
\title{Create a Spaghetti Plot From ws_monitor Data}
\usage{
gridPlot_spaghetti(bs_grid, longitude = NULL, latitude = NULL,
  count = 9, width = 3, ...)
}
\arguments{
\item{bs_grid}{a bs_grid list}

\item{longitude}{target longitude}

\item{latitude}{target latitude}

\item{count}{count of closest grid cells to be included in the plot}

\item{width}{rolling mean width in hours}

\item{...}{additional arguments to pass to plot() via monitorPlot_timeseries()}
}
\description{
Creates a "spaghetti plot" timeseries  with a line for each of the
\code{count} model grid cells closest to the target location. A centered
rolling mean is applied to each timeseries before plotting.
}
\examples{
\dontrun{
setModelDataDir('~/Data/Bluesky')
bluesky <- bluesky_aggregate(model = 'PNW-4km', firstModelRun = 20190915, lastModelRun = 20190917)
gridPlot_spaghetti(bluesky, longitude=-117.400, latitude=47.660)
}
}
\keyword{bs_grid,}
\keyword{spaghettiPlot}
