% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirFireModeling.R
\name{inside_model}
\alias{inside_model}
\title{Find coordinates appropriate model}
\usage{
inside_model(longitude, latitude)
}
\arguments{
\item{longitude}{the target longitude}

\item{latitude}{the target latitude}
}
\value{
vectors of model(s)
}
\description{
Determine what model to use based on the target coordinates
provided.
}
