% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirFireModeling.R
\name{load_check}
\alias{load_check}
\title{Show loading for futures}
\usage{
load_check(f, msg, verbose)
}
\arguments{
\item{f}{a future}

\item{msg}{a message to display}

\item{verbose}{logical. to display}
}
\description{
Show loading for futures
}
\keyword{Internal}
