% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirFireModeling.R
\name{modelBoundingBox}
\alias{modelBoundingBox}
\title{Find bs_grid model bouding box}
\usage{
modelBoundingBox(bs_grid)
}
\arguments{
\item{bs_grid}{a bs_grid object}
}
\value{
a bounding box matrix
}
\description{
Quickly find the bluesky models bounded region.
}
