% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirFireModeling.R
\name{modelInfo}
\alias{modelInfo}
\title{Model information}
\usage{
modelInfo(model)
}
\arguments{
\item{model}{the model to view info of.}
}
\description{
Used to find out model information, such as region and sub-
directory structure.
}
