% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_forecast.R
\name{monitor_forecast}
\alias{monitor_forecast}
\title{Compare forecasts}
\usage{
monitor_forecast(
  ws_monitor,
  starttime = NULL,
  endtime = NULL,
  models = c("CANSAC-1.33km", "CANSAC-4km"),
  subDir = "forecast",
  radius = 5
)
}
\arguments{
\item{ws_monitor}{a ws_monitor object.}

\item{starttime}{an optional start time to subset with}

\item{endtime}{an optional end time to subset with}

\item{models}{a list of valid monitors to plot}

\item{subDir}{Subdirectory path containing netcdf data. (Passed to
\code{raster_load()})}

\item{radius}{a radial buffer around the selected monitor to collapse the
model about.}
}
\description{
Overlays models of the bluesky forecasts to compare with ggplot.
}
