% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_ggforecastPlot.R
\name{monitor_ggforecastPlot}
\alias{monitor_ggforecastPlot}
\title{Compare forecasts}
\usage{
monitor_ggforecastPlot(ws_monitor, starttime = NULL, endtime = NULL,
  models = c("CANSAC-1.33km", "CANSAC-4km"), subDir = "forecast")
}
\arguments{
\item{ws_monitor}{a ws_monitor object.}

\item{starttime}{an optional start time to subset with}

\item{endtime}{an optional end time to subset with}

\item{models}{a list of valid monitors to plot}

\item{subDir}{Subdirectory path containing netcdf data. (Passed to
\code{bluesky_load()})}
}
\description{
Overlays models of the bluesky forecasts to compare with ggplot.
}
\examples{
\dontrun{
library(PWFSLSmoke)
initializeMazamaSpatialUtils()

library(AirFireModeling)
setModelDataDir('~/Data/Bluesky')

yosemite_village <-
  monitor_load(20191025, 20191029) \%>\%
  monitor_subset(monitorIDs = '060131004_01')

monitor_ggforecastPlot(yosemite_village)
}
}
