% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_toRaster.R
\name{monitor_toRaster}
\alias{monitor_toRaster}
\title{Create a Raster from monitor object}
\usage{
monitor_toRaster(
  ws_monitor,
  res,
  crs = "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0",
  extent = NULL,
  projectTo = NULL
)
}
\arguments{
\item{ws_monitor}{A ws_monitor object.}

\item{res}{Set the resolution of a Raster* object}

\item{crs}{Set the coordinate reference system (CRS) of a Raster* object.}

\item{extent}{Set or inherit the extent of the Raster* object.}

\item{projectTo}{Project the values of a Raster* object to a different Raster.}
}
\value{
a \code{raster} object.
}
\description{
Create a Raster from monitor object
}
