% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_aggregate.R
\name{raster_aggregate}
\alias{raster_aggregate}
\title{Aggregate Raster Models}
\usage{
raster_aggregate(
  model = "PNW-4km",
  startdate,
  enddate,
  by = c(1, 12),
  xlim = NULL,
  ylim = NULL,
  local = NULL,
  dirURL = "https://haze.airfire.org/bluesky-daily/output/standard",
  type = "forecast",
  clean = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{model}{Model identifier(s).}

\item{startdate}{Desired start date (integer or character in ymd[hms] format or POSIXct).}

\item{enddate}{Desired end date (integer or character in ymd[hms] format or POSIXct).}

\item{by}{Numeric vector. Represents hours to truncate models at.}

\item{xlim}{Optional longitude range.}

\item{ylim}{Optional latitude range.}

\item{local}{Absolute or relative paths of the local NetCDF (.nc) files.}

\item{dirURL}{Model output web directory. Default is BlueSky standard output.}

\item{type}{Model type directory, i.e. 'forecast', 'combined', etc.}

\item{clean}{Logical specifying whether to remove the non-common format NetCDF.}

\item{verbose}{If \code{FALSE}, suppress status messages (if any).}
}
\value{
A raster object.
}
\description{
Aggregate Raster Models
}
\examples{
\dontrun{

}
}
