% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_calculate.R
\name{raster_calculate}
\alias{raster_calculate}
\title{Calculate Values}
\usage{
raster_calculate(raster, FUN)
}
\arguments{
\item{raster}{A Raster\* object or a list of Raster\* objects.}

\item{FUN}{A function to apply to each RasterLayer (i.e. model time)}
}
\value{
A Raster* object
}
\description{
A wrapper around \code{raster::calc()}. Preform calculations for a
new Raster* object from another raster using a formula.
}
