% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-raster.R
\name{raster_copyfn}
\alias{raster_copyfn}
\title{Inherit File path across Raster objects}
\usage{
raster_copyfn(from, to, ...)
}
\arguments{
\item{from}{The raster to copy path from.}

\item{to}{the raster to copy path to.}
}
\value{
raster
}
\description{
Inherit File path across Raster objects
}
\keyword{internal}
