% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_correlation.R
\name{raster_correlation}
\alias{raster_correlation}
\title{Local correlation coefficient}
\usage{
raster_correlation(
  x,
  y,
  index = 1,
  plot = TRUE,
  res = NULL,
  method = "kendall",
  p_value = FALSE,
  crs = "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0"
)
}
\arguments{
\item{x}{A Raster* object}

\item{y}{A Raster* object}

\item{index}{A ratser index to access different layers}

\item{plot}{Logical. To plot or not.}

\item{res}{An optional resolution to project both x, y to.}

\item{method}{Correlation method. Default: 'kendall'.}

\item{p_value}{Logical. If TRUE, return a p-value}

\item{crs}{Set the coordinate reference system.}
}
\value{
a Raster* object
}
\description{
Local correlation coefficient
}
