% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-raster.R
\name{raster_createLayerNameTimes}
\alias{raster_createLayerNameTimes}
\title{Create raster layer name times}
\usage{
raster_createLayerNameTimes(layerName)
}
\arguments{
\item{layerName}{Vector of raster layer names.}
}
\value{
\code{POSIXct} time.
}
\description{
Create raster layer name times
}
\examples{
\donttest{
library(AirFireModeling)
setModelDataDir('~/Data/BlueSky')

# Load model data
rasterList <- raster_load(
  modelName = c("PNW-4km"),
  modelRun = c(2019100900),
  xlim = c(-125, -115),
  ylim = c(42, 50)
)

layerNames <- names(rasterList[[1]])
raster_createLayerNameTimes(layerNames)
}
}
\keyword{internal}
