% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_createMonitor.R
\name{raster_createMonitor}
\alias{raster_createMonitor}
\title{Create a ws_monitor object from raster object}
\usage{
raster_createMonitor(raster, longitude = NULL, latitude = NULL,
  buffer = 1000, monitorID = NULL, FUN = mean)
}
\arguments{
\item{raster}{A raster object}

\item{longitude}{Target longitude}

\item{latitude}{Target latitude}

\item{buffer}{A radial buffer about the buffer, in meters.}

\item{monitorID}{An optional monitor identification name.}

\item{FUN}{A function to collapse cells if buffer > 0.}
}
\value{
A \emph{ws_monitor} object representing a single monitor.
}
\description{
Time series associated with multiple grid cells are merged into a single
time series by using \code{FUN} to collapse a given grid cell count, or
or a grid cell radi, to a single coordinate. For instance, if the
\code{FUN = mean} then the grid cells within the paramters are averaged to a
single central coordinate.
}
