% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-raster.R
\name{raster_createTimeStamps}
\alias{raster_createTimeStamps}
\title{Create raster layer timestamp}
\usage{
raster_createTimeStamps(raster = NULL)
}
\arguments{
\item{raster}{A Raster\* object.}

\item{timezone}{Olson timezone in which times will be displayed.}

\item{prefix}{String prepended to the time}
}
\value{
UTC timestamp as YYYYmmddHH.
}
\description{
Create raster layer timestamp
}
\keyword{internal}
