% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_facet.R
\name{raster_facet}
\alias{raster_facet}
\title{Small Multiples of Raster}
\usage{
raster_facet(
  raster,
  breaks = "default",
  palette = "default",
  direction = 1,
  timezone = "UTC"
)
}
\arguments{
\item{raster}{a Raster* object}

\item{breaks}{Color break locations.}

\item{palette}{A color brewer palette (i.e. 'Greys', 'Blues', 'BuGn', 'YlOrRd', ...).}

\item{direction}{Color palette direction. Set -1 to reverse direction.}

\item{timezone}{Timezone.}
}
\value{
a gg object
}
\description{
Small Multiples of Raster
}
