% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-raster.R
\name{raster_generateTime}
\alias{raster_generateTime}
\title{Create raster layer times}
\usage{
raster_generateTime(raster = NULL)
}
\arguments{
\item{raster}{A RasterBrick.}
}
\value{
\code{POSIXct} time.
}
\description{
Create raster layer times
}
\examples{
\donttest{
library(AirFireModeling)
setModelDataDir('~/Data/BlueSky')

# Load model data
rasterList <- raster_load(
  modelName = c("PNW-4km"),
  modelRun = c(2019100900),
  xlim = c(-125, -115),
  ylim = c(42, 50)
)

raster_generateTime(rasterList[[1]])
}
}
\keyword{internal}
