% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-raster.R
\name{raster_getTime}
\alias{raster_getTime}
\title{Get the time-stamp(s) of a Raster object}
\usage{
raster_getTime(raster, tz = "UTC")
}
\arguments{
\item{raster}{A Raster* object.}

\item{tz}{A desired timezone. Default \code{'UTC'}.}
}
\value{
\code{POSIXct} vector
}
\description{
Get the time-stamp(s) of a raster object via the Raster* object
layer names in \code{POSIXct} format and UTC timezone.
}
