% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_map.R
\name{raster_ggmap}
\alias{raster_ggmap}
\alias{raster_ggmap.Raster}
\alias{raster_ggmap.list}
\title{Geographic map of BlueSky model output}
\usage{
raster_ggmap(
  raster,
  index = 1,
  palette = "Greys",
  breaks = c(0, 12, 35, 55, 150, 250, 350, Inf),
  direction = 1,
  title = "PM2.5",
  timezone = "UTC",
  col_state = "black",
  col_county = "gray80",
  verbose = TRUE
)

\method{raster_ggmap}{Raster}(
  raster,
  index = 1,
  palette = "Greys",
  breaks = c(0, 12, 35, 55, 150, 250, 350, Inf),
  direction = 1,
  title = "PM2.5",
  timezone = "UTC",
  col_state = "black",
  col_county = "gray80",
  verbose = TRUE
)

\method{raster_ggmap}{list}(
  raster,
  index = 1,
  palette = "Greys",
  breaks = c(0, 12, 35, 55, 150, 250, 350, Inf),
  direction = 1,
  title = "PM2.5",
  timezone = "UTC",
  col_state = "black",
  col_county = "gray80",
  verbose = TRUE
)
}
\arguments{
\item{raster}{A Raster\* object or a list of Raster\* objects.}

\item{index}{And index into the Raster\* object. See details.}

\item{palette}{The color palette used to map cell values. This must be one
of the palettes available through \code{ggplot2::scale_colour_brewer()}.}

\item{breaks}{The breaks used to map cell values to colors.}

\item{direction}{Numeric. \code{direction = -1} reverses color palette.}

\item{title}{(Optional) A plot title.}

\item{timezone}{Olson timezone in which times will be displayed.}

\item{col_state}{Color of state lines. (use \code{'transparent'} to hide them.)}

\item{col_county}{Color of county lines. (use \code{'transparent'} to hide them.)}

\item{verbose}{Logical to display messages.}
}
\value{
A ggplot object.
}
\description{
Creates a \pkg{ggplot2} plot object from a Raster\*
object. The returned plot object can be plotted or can be enhanced with
additional \pkg{ggplot2} intructions.
}
\details{
If a list of Raster\* objects is provided, a small-multiples
plot is created.

The \code{index} is typically associated with a time-axis or \code{RasterLayer},
e.g \code{index = 1} is the first hour of a model.
}
\section{Methods (by class)}{
\itemize{
\item \code{Raster}: Method for Raster* objects.

\item \code{list}: Multi-threaded method for a list of Raster* objects.
}}

\examples{
\donttest{
library(AirFireModeling)
setModelDataDir('~/Data/BlueSky')

# Load model data
rasterList <- raster_load(
  modelName = "PNW-4km",
  modelRun = c(2021010800, 2021010900, 2021011000, 2021011100),
  xlim = c(-125, -115),
  ylim = c(42, 50)
)

raster_ggmap(rasterList, title="PNW-4km", index = 3)
}
}
