% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_loadNearestMonitor.R
\name{raster_loadNearestMonitor}
\alias{raster_loadNearestMonitor}
\title{Title}
\usage{
raster_loadNearestMonitor(raster, longitude, latitude)
}
\arguments{
\item{latitude}{}
}
\value{

}
\description{
Title
}
