% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_map.R
\name{raster_map}
\alias{raster_map}
\alias{raster_map.Raster}
\alias{raster_map.list}
\title{Geographic map of BlueSky model output}
\usage{
raster_map(
  raster,
  index = 1,
  palette = "Greys",
  breaks = c(0, 12, 35, 55, 150, 250, 350, Inf),
  direction = 1,
  title = "PM2.5",
  timezone = "UTC",
  col_state = "black",
  col_county = "gray80",
  verbose = TRUE
)

\method{raster_map}{Raster}(
  raster,
  index = 1,
  palette = "Greys",
  breaks = c(0, 12, 35, 55, 150, 250, 350, 500),
  direction = 1,
  title = "PM2.5",
  timezone = "UTC",
  col_state = "black",
  col_county = "gray80",
  verbose = TRUE
)

\method{raster_map}{list}(
  raster,
  index = 1,
  palette = "Greys",
  breaks = c(0, 12, 35, 55, 150, 250, 350, 500),
  direction = 1,
  title = "PM2.5",
  timezone = "UTC",
  col_state = "black",
  col_county = "gray80",
  verbose = TRUE
)
}
\arguments{
\item{raster}{A \code{Raster\*} object or a list of \code{Raster\*} objects.}

\item{index}{And index into the \code{Raster*} object. See details.}

\item{palette}{The color palette used to map cell values. This must be one
of the palettes available through \code{ggplot2::scale_colour_brewer()}.}

\item{breaks}{The breaks used to map cell values to colors.}

\item{direction}{Numeric. \code{direction = -1} reverses color palette.}

\item{title}{(Optional) A plot title.}

\item{timezone}{Olson timezone in which times will be displayed.}

\item{col_state}{Color of state lines. (use \code{'transparent'} to hide them.)}

\item{col_county}{Color of county lines. (use \code{'transparent'} to hide them.)}

\item{verbose}{Logical to display messages.}
}
\value{
A ggplot object.
}
\description{
Creates a \pkg{ggplot2} plot object from a \code{Raster\*}
object. The returned plot object can be plotted or can be enhanced with
additional \pkg{ggplot2} intructions.
}
\details{
If a list of \code{Raster\*} objects is provided, a small-multiples
plot is created.

The \code{index} is typically associated with a time-axis or \code{RasterLayer},
e.g \code{index = 1} is the first hour of a model.
}
\section{Methods (by class)}{
\itemize{
\item \code{Raster}: Method for Raster* objects.

\item \code{list}: Multi-threaded method for a list of Raster* objects.
}}

\examples{
\dontrun{
library(AirFireModeling)
setModelDataDir('~/Data/BlueSky')

# Load from server
rasterList <- raster_load(
  model = "PNW-4km",
  modelRun = c(2019100800, 2019100900, 2019101000, 2019101100),
  xlim = c(-125, -115),
  ylim = c(42, 50)
)

raster_map(rasterList, title="PNW-4km", index = 3)
}
}
