% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_spaghetti.R
\name{raster_spaghetti}
\alias{raster_spaghetti}
\title{Spaghetti Plot}
\usage{
raster_spaghetti(raster, longitude = NULL, latitude = NULL, ...)
}
\arguments{
\item{raster}{A Raster* object}

\item{longitude}{A target latitude}

\item{latitude}{A target longitude}

\item{...}{See details.}
}
\value{
a gg object
}
\description{
Plot a spaghetti plot of all model adjacent cells to a target
location
}
\details{
\code{radius}: radial (meters) subset by distance from target location.
\code{n}: integer subset by the targets adjacent cell count.
}
