% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_spaghettiPlot.R
\name{raster_spaghettiPlot}
\alias{raster_spaghettiPlot}
\title{Spaghetti Plot}
\usage{
raster_spaghettiPlot(
  raster = NULL,
  longitude = NULL,
  latitude = NULL,
  radius = 50,
  count = NULL,
  monitorID = NULL
)
}
\arguments{
\item{raster}{A Raster\* object or a list of Raster\* objects.}

\item{longitude}{Target longitude from which the radius will be calculated.}

\item{latitude}{Target latitude from which the radius will be calculated.}

\item{radius}{Distance (km) of radius from target location.}

\item{count}{Number of grid cells within radius to return.}

\item{monitorID}{PWFSLSmoke monitorID used to retrieve and plot monitor data.}
}
\value{
A ggplot object.
}
\description{
Plot a spaghetti plot of all model adjacent cells to a target
location
}
\examples{

}
\seealso{
\code{\link{raster_subsetByDistance}}
}
