% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_subset.R
\name{raster_subset}
\alias{raster_subset}
\title{Subset a Raster* by distance from a location}
\usage{
raster_subset(raster, ...)
}
\arguments{
\item{raster}{A Raster* object.}

\item{...}{Subsetting arguments}
}
\value{
a Raster* object
}
\description{
Subset a Raster* by distance from a location
}
\examples{
\dontrun{
model <- raster_load()
pnw_4km_subbed <- raster_subset(model, n = 200, longitude = -120, latitude = 45)
}
}
