% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_subsetByDistance.R
\name{raster_subsetByDistance}
\alias{raster_subsetByDistance}
\title{Subset a Raster* by distance from a location}
\usage{
raster_subsetByDistance(
  raster = NULL,
  longitude = NULL,
  latitude = NULL,
  radius = 50,
  count = NULL
)
}
\arguments{
\item{raster}{A Raster* object.}

\item{longitude}{Target longitude from which the radius will be calculated.}

\item{latitude}{Target latitude from which the radius will be calculated.}

\item{radius}{Distance (km) of radius from target location.}

\item{count}{Number of grid cells within radius to return.}
}
\value{
A Raster* object.
}
\description{
This function subsets a raster object just like a round cookie
cutter. All cells (up to \code{count}) within \code{radius} km of
the target location are returned.
}
\examples{
\donttest{
library(AirFireModeling)
setModelDataDir('~/Data/BlueSky')

# Load model data
rasterList <- raster_load(
  modelName = "PNW-4km",
  modelRun = c(2020091300, 2020091400, 2020091500, 2020091600),
  xlim = c(-125, -115),
  ylim = c(42, 50)
)

Portland_area <- raster_subsetByDistance(
  rasterList,
  longitude = -122.68,
  latitude = 45.52,
  radius = 100
)

raster_ggmap(Portland_area)
}
}
