% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{roc}
\alias{roc}
\title{ROC and AUC}
\usage{
roc(predicted, observed, p_thresh = c(1, 100), o_thresh = 55.5,
  n = 100)
}
\arguments{
\item{predicted}{Predicted logical.}

\item{observed}{Observed logical.}

\item{p_thresh}{low and high values used to generate test thresholds for
classifying \code{predicted} data.}

\item{o_thresh}{value used to classify \code{observed} data.}

\item{n}{number of test thresholds in ROC curve.}
}
\value{
a AUROC list
}
\description{
ROC and AUC
}
