% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skill_confusionMatrix.R
\name{skill_confusionMatrix}
\alias{skill_confusionMatrix}
\title{Confusion Matrix}
\usage{
skill_confusionMatrix(predicted, observed)
}
\arguments{
\item{predicted}{Preicted logical.}

\item{observed}{Observed logical.}
}
\value{
a list of contingency statistics
}
\description{
Confusion Matrix
}
\examples{
\dontrun{
predicted <- sample(c(T,F), 100, replace = TRUE)
observed <- sample(c(T,F), 100, replace = TRUE)
skill_confusionMatrix(predicted, observed)
}
}
