% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skill_roc.R
\name{skill_roc}
\alias{skill_roc}
\title{ROC and AUC}
\usage{
skill_roc(predicted, observed, p_thresh = c(1, 100), o_thresh = 55.5, n = 100)
}
\arguments{
\item{predicted}{Predicted logical.}

\item{observed}{Observed logical.}

\item{p_thresh}{low and high values used to generate test thresholds for
classifying \code{predicted} data.}

\item{o_thresh}{value used to classify \code{observed} data.}

\item{n}{number of test thresholds in ROC curve.}
}
\value{
a AUROC list
}
\description{
ROC and AUC
}
\examples{
\dontrun{
ca <- airnow_loadAnnual(2017) \%>\%
  monitor_subset(tlim = c(20171001,20171101), stateCodes = 'CA')
Vallejo <- monitor_subset(ca, monitorIDs = '060950004_01')
Napa <- monitor_subset(ca, monitorIDs = '060550003_01')
skill_roc(predicted = Vallejo, observed = Napa, n = 10, o_thresh = 10, p_thresh = c(1,100))
}
}
