% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluesky_aggregate.R
\name{bluesky_aggregate}
\alias{bluesky_aggregate}
\title{Aggregate BlueSky models by hour}
\usage{
bluesky_aggregate(
  start,
  end,
  by = c(1, 12),
  model = "CANSAC-4km",
  type = "forecast",
  base_url = "https://haze.airfire.org/bluesky-daily/output",
  sub_dir = "standard",
  ...
)
}
\arguments{
\item{start}{A start date.}

\item{end}{A end date.}

\item{by}{Numeric vector. Represents hours to truncate models at.}

\item{model}{Model name (i.e. 'CANSAC-1.33km').}

\item{type}{Model type (i.e. 'forecast).}

\item{base_url}{The base url of the model output directory.}

\item{sub_dir}{The subdirectory of the output directory.}

\item{...}{TBD}
}
\value{
A Raster* brick

GFS-0.15deg-CanadaUSA-p25deg-68N'
GFS-0.15deg-CanadaUSA-p25deg-68N'
}
\description{
Aggregate BlueSky models by hour
}
