% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluesky_download.R
\name{bluesky_download}
\alias{bluesky_download}
\title{Download BlueSky model data from AirFire}
\usage{
bluesky_download(dailyOutputDir = "standard", model = "PNW-1.33km",
  modelRun = NULL, subDir = "combined",
  baseUrl = "https://haze.airfire.org/bluesky-daily/output",
  quiet = FALSE)
}
\arguments{
\item{dailyOutputDir}{BlueSky web directory}

\item{model}{Model identifier.}

\item{modelRun}{Date code as "YYYYMMDDHH".}

\item{subDir}{Subdirectory path containing netcdf data.}

\item{baseUrl}{Base URL for BlueSky output.}

\item{quiet}{If \code{TRUE}, suppress status messages (if any), and the
progress bar.}
}
\value{
File path of downloaded data.
}
\description{
Downloads a copy of the specified BlueSky model run to the 
package data directory. This file can then be converted into a more 
modernized format by bluesky2v2().

#' On 2019-10-11, vailable model identifiers include the following:
\itemize{
  \item{AK-12km}
  \item{CANSAC-1.33km}
  \item{CANSAC-4km}
  \item{DRI1.33km-CMAQ}
  \item{DRI4km-CMAQ}
  \item{GFS-0.15deg-CanadaUSA-p25deg-68N}
  \item{GFS-0.15deg}
  \item{NAM-3km}
  \item{NAM84-0.15deg}
  \item{PNW-1.33km}
  \item{PNW-4km}
  \item{PNW1.33km-CMAQ}
  \item{PNW4km-CMAQ}
}

Users will typically call bluesky_load() which in turn calls this function.


BlueSky output files are found in directories with the following
structure:

\preformatted{
.../standard/NAM84-0.08deg/2016050600/carryover/data/...
.../standard/NAM84-0.08deg/2016050600/combined/data/...
.../standard/NAM84-0.08deg/2016050600/forecast/data/...
}
}
\examples{
\dontrun{
setModelDataDir("~/Data/Bluesky")
bluesky_download(model = "PNW-4km", modelRun = 2019090900)
}
}
\seealso{
\link{bluesky_load}

\link{setModelDataDir}
}
