% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluesky_downloaded.R
\name{bluesky_downloaded}
\alias{bluesky_downloaded}
\title{List downloaded models}
\usage{
bluesky_downloaded(
  modelDataDir = getModelDataDir(),
  pattern = ".nc",
  full = FALSE,
  ...
)
}
\arguments{
\item{modelDataDir}{Directory in which models are located.}

\item{pattern}{A regex pattern to use for filtering model files.}

\item{full}{Logical. Show the full path of the model (used for local loading).}

\item{...}{Additional arguments to be passed to \code{list.files()}.}
}
\value{
A list of downloaded models
}
\description{
List downloaded models
}
\examples{
\donttest{
library(AirFireModeling)
setModelDataDir('~/Data/BlueSky')

filePath <- bluesky_download(modelName = "PNW-4km", modelRun = 2019100900)
bluesky_toCommonFormat(filePath)
bluesky_downloaded()
}

}
