% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluesky_load.R
\name{bluesky_load}
\alias{bluesky_load}
\title{Single model load function}
\usage{
bluesky_load(
  model = "PNW-4km",
  modelRun = NULL,
  baseUrl = "https://haze.airfire.org/bluesky-daily/output/standard",
  modelType = "forecast",
  localPath = NULL,
  xlim = NULL,
  ylim = NULL,
  clean = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{model}{A model. i.e 'PNW-1.33km' or 'CANSAC-4km'.}

\item{modelRun}{The model run date. YYYYmmddHH format required.}

\item{baseUrl}{the database URL.}

\item{modelType}{A type of model, formely subDir. i.e. 'forecast'}

\item{localPath}{A path to a downloaded NetCDF model.}

\item{xlim}{A vector of coordinate longitude bounds.}

\item{ylim}{A vector of coordinate latitude bounds.}

\item{clean}{Option to clean, or delete, the non-formatted model.}

\item{verbose}{Logical to display messages.}
}
\value{
A RasterBrick.
}
\description{
This function encapsulates the process of downloading, and formatting
necessary for BlueSky model outputs. Additionally, it checks already existing
models and loads those if avaliable.
}
\examples{
\dontrun{
library(AirFireModeling)

setModelDataDir('~/Data/BlueSky')
rasterBrick <- bluesky_load(
  model = "PNW-4km",
  modelRun = 2019100900,
  xlim = c(-125, -115),
  ylim = c(42, 50)
)

raster_map(rasterBrick, index = 3)
}

}
