% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluesky_load.R
\name{bluesky_load}
\alias{bluesky_load}
\title{Load data for a single BlueSky model}
\usage{
bluesky_load(
  modelName = NULL,
  modelRun = NULL,
  modelMode = "forecast",
  baseUrl = "https://haze.airfire.org/bluesky-daily/output/standard",
  localPath = NULL,
  level = 1,
  xlim = NULL,
  ylim = NULL,
  clean = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{modelName}{Model identifier.}

\item{modelRun}{Model initialization datestamp as "YYYYMMDDHH".}

\item{modelMode}{Subdirectory path containing BlueSky output, i.e. 'forcast'.}

\item{baseUrl}{Base URL for BlueSky output.}

\item{localPath}{Absolute path to a NetCDF file not found in `modelDataDir`.}

\item{level}{Elevation level.}

\item{xlim}{A vector of coordinate longitude bounds.}

\item{ylim}{A vector of coordinate latitude bounds.}

\item{clean}{Logical specifying removal of original model data after conversion
to "v2" format.}

\item{verbose}{Logical to display messages.}
}
\value{
A \pkg{raster} package \emph{RasterBrick} object.
}
\description{
This function encapsulates the process of downloading, and formatting
necessary for BlueSky model outputs. Additionally, it checks already existing
models and loads those if available.

Loads BlueSky model output as a raster brick. The directory
previously set with \code{setModelDataDir()} is searched for "v2" formatted
model output. If data are not found locally, data are downloaded from
\code{baseUrl} and converted.

The returned \code{dataBrick} object is of class \code{raster::RasterBrick} and
can be manipulated with appropriate functions from the \pkg{raster} package
or any of the \code{raster_~()} functions provided by \pkg{AirFireModeling}.

On 2019-10-11, available model identifiers include the following:
\itemize{
  \item{AK-12km}
  \item{CANSAC-1.33km}
  \item{CANSAC-4km}
  \item{DRI1.33km-CMAQ}
  \item{DRI4km-CMAQ}
  \item{GFS-0.15deg-CanadaUSA-p25deg-68N}
  \item{GFS-0.15deg}
  \item{NAM-3km}
  \item{NAM84-0.15deg}
  \item{PNW-1.33km}
  \item{PNW-4km}
  \item{PNW1.33km-CMAQ}
  \item{PNW4km-CMAQ}
}
}
\note{
The regurned \emph{RasterBrick} object will contain all grid cells
within the area defined by \code{xlim} and \code{ylim}.

If \code{localPath} is used, \code{clean = FALSE} is automatically set.
The package will never remove original model output that it has not downloaded.
}
\examples{
\donttest{
library(AirFireModeling)
setModelDataDir('~/Data/BlueSky')

# Load model data
raster <- bluesky_load(
  modelName = "PNW-4km",
  modelRun = 2021010800,
  xlim = c(-125, -115),
  ylim = c(42, 50)
)

raster_ggmap(raster, index = 3)
}
}
