% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_forecastPlot.R
\name{monitor_forecastPlot}
\alias{monitor_forecastPlot}
\title{Compare forecasts}
\usage{
monitor_forecastPlot(ws_monitor, starttime = NULL, endtime = NULL,
  models = c("CANSAC-1.33km", "CANSAC-4km"), subDir = "forecast",
  radius = 20000, count = 9, prob = 0.5, ...)
}
\arguments{
\item{ws_monitor}{a ws_monitor object.}

\item{starttime}{an optional start time to subset with}

\item{endtime}{an optional end time to subset with}

\item{models}{a list of valid monitors to plot}

\item{subDir}{Subdirectory path containing netcdf data. (Passed to}

\item{radius}{Radius (m) used to select model grid cells.}

\item{count}{Count of model grid cells to be used.}

\item{prob}{Quantile probability used when  plotting model data.
\code{bluesky_load()})}

\item{...}{Additional arguments passed to 
\code{PWFSLSmoke::monitor_timeseriesPlot()}.}
}
\description{
Plots monitor data and overlays models of the bluesky forecasts.

For each model, timeseries is created by gathering \code{count} grid cells
within \code{radius} meters of the \code{ws_monitor}. The resulting set of
timeseries are then collapsed into a single timeseries by applying the
\code{quantile()} using the user specified probability \code{prob}.
}
\examples{
\dontrun{
library(PWFSLSmoke)
initializeMazamaSpatialUtils()

library(AirFireModeling)
setModelDataDir('~/Data/Bluesky')

yosemite_village <-
  monitor_load(20191025, 20191029) \%>\%
  monitor_subset(monitorIDs = '060131004_01')

monitor_forecastPlot(yosemite_village)
}
}
