% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_forecastPlot.R
\name{monitor_forecastPlot}
\alias{monitor_forecastPlot}
\title{Compare forecasts}
\usage{
monitor_forecastPlot(ws_monitor, starttime = NULL, endtime = NULL,
  models = c("CANSAC-1.33km", "CANSAC-4km"), subDir = "combined")
}
\arguments{
\item{ws_monitor}{a ws_monitor object.}

\item{starttime}{an optional start time to subset with}

\item{endtime}{an optional end time to subset with}

\item{models}{a list of valid monitors to plot}

\item{subDir}{Subdirectory path containing netcdf data. (Passed to 
\code{bluesky_load()})}
}
\description{
Overlays models of the bluesky forecasts to compare.
}
\examples{
\dontrun{
library(PWFSLSmoke)
initializeMazamaSpatialUtils()

library(AirFireModeling)
setModelDataDir('~/Data/Bluesky')

pollock_pines <- 
  monitor_load(20191007, 20191014) \%>\%
  monitor_subset(monitorIDs = "lon_.120.591_lat_38.714_arb2.1008")
  
monitor_forecastPlot(pollock_pines)
}
}
