% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_toRaster.R
\name{monitor_toRaster}
\alias{monitor_toRaster}
\title{Create a Raster from monitor object}
\usage{
monitor_toRaster(ws_monitor = NULL, raster = NULL, FUN = mean)
}
\arguments{
\item{ws_monitor}{\pkg{PWFSLSMoke} \emph{ws_monitor} object.}

\item{raster}{Raster* object defining the output grid.}

\item{FUN}{Function used to collapse values when multiple monitors lie within
a single grid cell.}
}
\value{
A \code{RasterBrick} object.
}
\description{
Create a Raster from monitor object
}
\examples{
\donttest{
# library(AirFireModeling)
setModelDataDir('~/Data/BlueSky')

# Load model data
rasterList <- raster_load(
  modelName = c("PNW-4km"),
  modelRun = c(2019100900),
  xlim = c(-125, -117),
  ylim = c(45.5, 49)
)

# Load monitor data
ws_monitor <-
  PWFSLSmoke::monitor_load(
    startdate = 2019100901,
    enddate = 2019101223
  ) \%>\%
  PWFSLSmoke::monitor_subset(stateCodes = "WA")

monitor_rasterBrick <- monitor_toRaster(ws_monitor, rasterList[[1]], mean)

raster_facet(
  monitor_rasterBrick,
  index = 1:6,
  title = "Monitoring data",
  palette = 'Spectral',
  col_county = 'gray95',
  direction = -1,
  breaks = c(-Inf, 0, 12, 35, 55, 150, 250, 350, Inf)
)
}

}
