% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_toRaster.R
\name{monitor_toRaster}
\alias{monitor_toRaster}
\title{Create a Raster from monitor object}
\usage{
monitor_toRaster(ws_monitor = NULL, raster = NULL, FUN = mean)
}
\arguments{
\item{ws_monitor}{\pkg{PWFSLSMoke} \emph{ws_monitor} object.}

\item{raster}{Raster* object defining the output grid.}

\item{FUN}{Function used to collapse values when multiple monitors lie within
a single grid cell.}
}
\value{
A \code{RasterBrick} object.
}
\description{
Create a Raster from monitor object
}
