% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_coordinateTrace.R
\name{raster_coordinateTrace}
\alias{raster_coordinateTrace}
\alias{raster_coordinateTrace.list}
\alias{raster_coordinateTrace.Raster}
\title{Plot a raster coordinate-temporal values}
\usage{
raster_coordinateTrace(
  raster,
  longitude = NULL,
  latitude = NULL,
  monitorID = NULL,
  tlim = "default",
  ...
)

\method{raster_coordinateTrace}{list}(
  raster,
  longitude = NULL,
  latitude = NULL,
  monitorID = NULL,
  tlim = "default",
  ...
)

\method{raster_coordinateTrace}{Raster}(raster, longitude, latitude, monitorID = NULL, ...)
}
\arguments{
\item{raster}{A Raster\* object or a list of Raster\* objects.}

\item{longitude}{(Optional) Target longitude to obtain values.}

\item{latitude}{(Optional) Target latitude to obtain values.}

\item{monitorID}{(Optional) a monitor ID to load}

\item{tlim}{(Optional) a time limit to restrict the plot axis.}

\item{...}{additional parameters. See details.}
}
\value{
A ggplot object
}
\description{
Plot a raster coordinate-temporal values
}
\section{Methods (by class)}{
\itemize{
\item \code{list}: A multithreaded implementation for lists of Raster* objects.

\item \code{Raster}: A Raster* object implementation.
}}

