% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-raster.R
\name{raster_createLayerNameTimes}
\alias{raster_createLayerNameTimes}
\title{Create raster layer name times}
\usage{
raster_createLayerNameTimes(layerName)
}
\arguments{
\item{layerName}{Vector of raster layer names.}
}
\value{
\code{POSIXct} time.
}
\description{
Create raster layer name times
}
\keyword{internal}
