% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-raster.R
\name{raster_createTimes}
\alias{raster_createTimes}
\title{Create raster layer times}
\usage{
raster_createTimes(raster = NULL)
}
\arguments{
\item{raster}{A Raster\* object.}
}
\value{
\code{POSIXct} time.
}
\description{
Create raster layer times
}
\keyword{internal}
