% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-raster.R
\name{raster_createTimeStrings}
\alias{raster_createTimeStrings}
\title{Create raster layer time string}
\usage{
raster_createTimeStrings(
  raster = NULL,
  format = "\%Y-\%m-\%d \%H:00 \%Z",
  timezone = "UTC",
  prefix = ""
)
}
\arguments{
\item{raster}{A RasterBrick.}

\item{format}{Format passed on to \code{strftime()}.}

\item{timezone}{Olson timezone in which times will be displayed.}

\item{prefix}{String prepended to the time.}
}
\value{
Time string formatted for \code{timezone}.
}
\description{
Create raster layer time string
}
\examples{
\donttest{
library(AirFireModeling)
setModelDataDir('~/Data/BlueSky')

# Load model data
rasterList <- raster_load(
  modelName = c("PNW-4km"),
  modelRun = c(2019100900),
  xlim = c(-125, -115),
  ylim = c(42, 50)
)

raster_createTimeStrings(
  raster = rasterList[[1]],
  timezone = "America/Los_Angeles",
  prefix = "time: ")
}
}
