% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-raster.R
\name{raster_createTimeStrings}
\alias{raster_createTimeStrings}
\title{Create raster layer time string}
\usage{
raster_createTimeStrings(raster = NULL, timezone = "UTC", prefix = "")
}
\arguments{
\item{raster}{A Raster\* object.}

\item{timezone}{Olson timezone in which times will be displayed.}

\item{prefix}{String prepended to the time}
}
\value{
Time string formatted for \code{timezone}.
}
\description{
Create raster layer time string
}
