% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_facet.R
\name{raster_facet}
\alias{raster_facet}
\title{Small multiples map of BlueSky model output}
\usage{
raster_facet(
  raster,
  index = NULL,
  palette = "Greys",
  breaks = c(0, 12, 35, 55, 150, 250, 350, Inf),
  direction = 1,
  title = "",
  timezone = "UTC",
  ncol = NULL,
  col_state = "black",
  col_county = "gray80"
)
}
\arguments{
\item{raster}{A Raster\* object.}

\item{index}{Vector of indices used to subset \code{raster}}

\item{palette}{Color palette used to map cell values. This must be either
\code{"aqi"} or one of the palette names available through
\code{ggplot2::scale_colour_brewer()}.}

\item{breaks}{The breaks used to map cell values to colors. (Unless
\code{palette = "aqi"} in which case AQI breaks are used.)}

\item{direction}{Numeric. \code{direction = -1} reverses color palette.}

\item{title}{(Optional) A plot title.}

\item{timezone}{Olson timezone in which times will be displayed.}

\item{ncol}{Number of columns in the output plot.}

\item{col_state}{Color of state lines. (use \code{'transparent'} to hide them.)}

\item{col_county}{Color of county lines. (use \code{'transparent'} to hide them.)}
}
\value{
A gg object.
}
\description{
This function creates a grid of plots, each displaying a small
map of a single hour from the model output passed in as \code{raster}. By
default, all hours are plotted. The \code{index} parameter can be used to
subset the hours.
}
\examples{
\dontrun{
library(AirFireModeling)
setModelDataDir('~/Data/BlueSky')

# Load model data
rasterList <- raster_load(
  model = "PNW-4km",
  modelRun = c(2019100900),
  xlim = c(-125, -117),
  ylim = c(42, 47)
)

# First 24 hours
raster_facet(
  rasterList[[1]],
  index = 1:24,
  col_county = 'transparent'
)

# Fancy!!
raster_facet(
  rasterList[[1]],
  index = seq(3,72,12),
  title = "PNW-4km -- run 2019100900",
  timezone = "America/Los_Angeles"
)

# Kincade fire
rasterList <- raster_load(
  model = "CANSAC-4km",
  modelRun = c(2019102700),
  xlim = c(-124, -121.5),
  ylim = c(37.5, 39)
)

raster_facet(
  rasterList[[1]],
  title = "Kincade Fire",
  ncol = 12
)

}

}
