% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-raster.R
\name{raster_getTime}
\alias{raster_getTime}
\title{Get the time-stamp(s) of a Raster object}
\usage{
raster_getTime(raster)
}
\arguments{
\item{raster}{A Raster* object.}
}
\value{
POSIXct vector
}
\description{
Get the time-stamp(s) of a Raster object
}
