% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-raster.R
\name{raster_isRaster}
\alias{raster_isRaster}
\title{Test for Raster\* class}
\usage{
raster_isRaster(raster)
}
\arguments{
\item{raster}{Object to test.}
}
\value{
\code{TRUE} if \code{raster} is a Raster\* object, \code{FALSE} otherwise.
}
\description{
Test for Raster\* class
}
\examples{
\donttest{
library(AirFireModeling)
setModelDataDir('~/Data/BlueSky')

# Load model data
rasterList <- raster_load(
  model = c("PNW-4km"),
  modelRun = c(2019100900),
  xlim = c(-125, -115),
  ylim = c(42, 50)
)

raster_isRaster(rasterList[[1]])
}

}
\keyword{hidden}
