% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_leaflet.R
\name{raster_leaflet}
\alias{raster_leaflet}
\title{Raster Leaflet Map}
\usage{
raster_leaflet(raster, compare = NULL, palette = "Spectral", direction = -1)
}
\arguments{
\item{raster}{A Raster* object}

\item{compare}{An optional comparison Raster}

\item{palette}{A color palette, default is Spectral}

\item{direction}{Color palette direction. -1 is reverse.}
}
\value{
a leaflet map
}
\description{
Raster Leaflet Map
}
