% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_leaflet.R
\name{raster_leaflet}
\alias{raster_leaflet}
\title{Raster Leaflet Map}
\usage{
raster_leaflet(
  raster,
  index = 1,
  palette = "Spectral",
  breaks = "default",
  title = "PM2.5",
  direction = -1,
  timezone = "UTC",
  compare = NULL,
  index2 = 1
)
}
\arguments{
\item{raster}{A \code{Raster\*} object.}

\item{index}{The index of the \code{Raster*} object. See details.}

\item{palette}{The color palette used to map cell values. This must be one
of the palettes available through \code{ggplot2::scale_colour_brewer()}.}

\item{breaks}{The breaks used to define the map cell color or \code{'default'}. CURRENTLY UNUSED}

\item{title}{(Optional) A plot title. CURRENTLY UNUSED}

\item{direction}{Numeric. \code{direction = -1} reverses color palette.}

\item{timezone}{The timezone to parse the dates. Default is UTC.}

\item{compare}{An optional comparison \code{Raster*} object}

\item{index2}{The index of the comparison \code{Raster*} object.}
}
\value{
a leaflet map
}
\description{
Create a leaflet interactive map using a model Raster Image.
}
\details{
The \code{index} is typically associated with a time-axis or RasterLayer,
e.g \code{index = 1} is the first hour of a model.
}
\examples{
\dontrun{
library(AirFireModeling)
setModelDataDir('~/Data/BlueSky')

# Load consecutive model runs
rasterList <- raster_load(
  model = "PNW-4km",
  modelRun = c(2019100900, 2019101000),
  xlim = c(-125, -117),
  ylim = c(42, 47)
)

# Create a leaflet map
raster_leaflet(
  raster = rasterList[[1]],
  index = 3,
  title = "PNW-4km"
)

# Create a leaflet map comparing two times from the same run
raster_leaflet(
  raster = rasterList[[1]],
  index = 15,
  compare = rasterList[[1]],
  index2 = 18,
  title = "PNW-4km",
  timezone = "America/Los_Angeles"
)

# Create a leaflet map comparing the same time from two different runs
raster_leaflet(
  raster = rasterList[[1]],
  index = 27,
  compare = rasterList[[2]],
  index2 = 3,
  title = "PNW-4km",
  timezone = "America/Los_Angeles"
  )
}
}
