% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_leaflet.R
\name{raster_leaflet}
\alias{raster_leaflet}
\title{Raster Leaflet Map}
\usage{
raster_leaflet(
  raster,
  index = 1,
  palette = "Spectral",
  breaks = c(0, 12, 35, 55, 150, 250, 350, Inf),
  direction = -1,
  opacity = 0.6,
  maptype = "terrain",
  title = "PM2.5",
  timezone = "UTC",
  compare = NULL,
  index2 = 1
)
}
\arguments{
\item{raster}{A Raster\* object.}

\item{index}{And index into the Raster\* object. See details.}

\item{palette}{Color palette used to map cell values. This must be either
\code{"aqi"} or one of the palette names available through
\code{ggplot2::scale_colour_brewer()}.}

\item{breaks}{The breaks used to map cell values to colors. CURRENTLY UNUSED}

\item{direction}{Numeric. \code{direction = -1} reverses color palette.}

\item{opacity}{Opacity of raster layer.}

\item{maptype}{Name of leaflet ProviderTiles to use.}

\item{title}{(Optional) A plot title. CURRENTLY UNUSED}

\item{timezone}{Olson timezone in which times will be displayed.}

\item{compare}{Optional comparison Raster\* object.}

\item{index2}{Index of the comparison Raster\* object.}
}
\value{
A leaflet map.
}
\description{
Create a leaflet interactive map using a model output
Raster\* object.
}
\details{
The \code{index} is associated with the z axis or time-axis of
\code{RasterBrick} object, e.g \code{index = 1} is the first hour of a model.

The \code{maptype} argument is mapped onto leaflet "ProviderTile" names.
Current mappings include:
\enumerate{
\item{"roadmap"}{ -- "OpenStreetMap"}
\item{"satellite"}{ -- "Esri.WorldImagery"}
\item{"terrain"}{ -- "Esri.WorldTopoMap"}
\item{"toner"}{ -- "Stamen.Toner"}
}

If a character string not listed above is provided, it will be used as the
underlying map tile if available. See
\url{https://leaflet-extras.github.io/leaflet-providers/} for a list of
"provider tiles" to use as the background map.
}
\examples{
\dontrun{
library(AirFireModeling)
setModelDataDir('~/Data/BlueSky')

# Load consecutive model runs
rasterList <- raster_load(
  modelName = "PNW-4km",
  modelRun = c(2019100900, 2019101000),
  xlim = c(-125, -117),
  ylim = c(42, 47)
)

# Create a leaflet map
raster_leaflet(
  raster = rasterList[[1]],
  index = 3
)

# Create a leaflet map comparing two times from the same run
raster_leaflet(
  raster = rasterList[[1]],
  index = 15,
  compare = rasterList[[1]],
  index2 = 18,
  timezone = "America/Los_Angeles"
)

# Create a leaflet map comparing the same time from two different runs

# Load consecutive model runs
rasterList <- raster_load(
  modelName = c("PNW-4km", "PNW-1.33km"),
  modelRun = 2019100900,
  xlim = c(-125, -117),
  ylim = c(42, 47)
)

raster_leaflet(
  raster = rasterList[[1]],
  index = 15,
  compare = rasterList[[2]],
  index2 = 15,
  timezone = "America/Los_Angeles",
  opacity = 0.3
  )
}
}
