% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_load.R
\name{raster_load}
\alias{raster_load}
\title{Load/Download a Raster Model}
\usage{
raster_load(
  model = "PNW-4km",
  run = NULL,
  xlim = NULL,
  ylim = NULL,
  local = NULL,
  dirURL = "https://haze.airfire.org/bluesky-daily/output/standard",
  type = "forecast",
  clean = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{model}{Model identifier(s).}

\item{run}{Date code as "YYYYMMDDHH" (integer or character).}

\item{xlim}{Optional longitude range.}

\item{ylim}{Optional latitude range.}

\item{local}{Absolute path of the local NetCDF (.nc) file}

\item{dirURL}{Model output web directory. Default is BlueSky standard output.}

\item{type}{Model type directory, i.e. 'forecast', 'combined', etc.}

\item{clean}{Logical specifying whether to remove the non-common format NetCDF.}

\item{verbose}{If \code{FALSE}, suppress status messages (if any).}
}
\value{
A raster object, or a list of raster objects
}
\description{
Load or download models and automatically convert NetCDF (.nc) format to
a spatio-temporal \code{Raster*} object.
}
\details{
\code{model}, \code{run}, and \code{local} are all vectorised
parameters. As such, loading a model is executed using multiple CPU threads.
}
\examples{
\dontrun{
# Set Model download directory (Required)
setModelDataDir('~/Data/Bluesky')
# Load from server
model <- raster_load('PNW-1.33km', run = 20200303, xlim = c(-125, -115))
# Load from local NetCDF
model <- raster_load('~/Data/Bluesky/PNW-4km_2020030100.nc')
}
}
