% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_load.R
\name{raster_load}
\alias{raster_load}
\title{Load BlueSky model data}
\usage{
raster_load(
  model = NULL,
  modelRun = NULL,
  modelType = "forecast",
  baseUrl = "https://haze.airfire.org/bluesky-daily/output/standard",
  localPath = NULL,
  xlim = NULL,
  ylim = NULL,
  clean = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{model}{Model identifier.}

\item{modelRun}{Model initialization datestamp as "YYYYMMDDHH".}

\item{modelType}{Subdirectory path containing BlueSky output, i.e. 'forcast'.}

\item{baseUrl}{Base URL for BlueSky output.}

\item{localPath}{Vector of absolute paths to NetCDF files not found in `modelDataDir`.}

\item{xlim}{Vector of coordinate longitude bounds.}

\item{ylim}{Vector of coordinate latitude bounds.}

\item{clean}{Logical specifying removal of original model data after conversion
to "v2" format.}

\item{verbose}{Logical to display messages.}
}
\value{
A list of one or more raster objects.
}
\description{
Load or download models and automatically convert NetCDF (.nc)
format to a spatio-temporal Raster\* object.
}
\details{
\code{model} and \code{modelRun} are vectorised parameters. Multiple
datasets can be loaded at once.
}
\examples{
\donttest{
library(AirFireModeling)
setModelDataDir('~/Data/BlueSky')

# Load model data
rasterList <- raster_load(
  model = "PNW-4km",
  modelRun = c(2019100800, 2019100900, 2019101000, 2019101100),
  xlim = c(-125, -115),
  ylim = c(42, 50)
)

raster_ggmap(rasterList, index = 3)
}
}
