% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_loadNearestMonitor.R
\name{raster_loadNearestMonitor}
\alias{raster_loadNearestMonitor}
\title{Load nearest monitor to coordinate location}
\usage{
raster_loadNearestMonitor(raster, longitude, latitude)
}
\arguments{
\item{raster}{A raster* object}

\item{longitude}{A latitude}

\item{latitude}{A longitude}
}
\value{
a ws_onitor
}
\description{
Load nearest monitor to coordinate location
}
