% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_map.R
\name{raster_map}
\alias{raster_map}
\title{Plot a Raster Map}
\usage{
raster_map(raster, palette = "Greys", breaks = "default", title = "PM2.5")
}
\arguments{
\item{raster}{A Raster* object.}

\item{palette}{A color brewer palette (i.e. 'Greys', 'Blues', 'BuGn', 'YlOrRd', ...).}

\item{breaks}{Color break locations.}

\item{title}{A plot title.}
}
\value{
a ggobject
}
\description{
Plot a Raster Map
}
