% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_map.R
\name{raster_map}
\alias{raster_map}
\alias{raster_map.Raster}
\alias{raster_map.list}
\title{Plot maps of BlueSky model output}
\usage{
raster_map(
  raster,
  index = 1,
  palette = "Greys",
  breaks = "default",
  title = "PM2.5",
  direction = 1,
  timezone = "UTC",
  verbose = TRUE
)

\method{raster_map}{Raster}(
  raster,
  index = 1,
  palette = "Greys",
  breaks = "default",
  title = "PM2.5",
  direction = 1,
  timezone = "UTC",
  verbose = TRUE
)

\method{raster_map}{list}(
  raster,
  index = 1,
  palette = "Greys",
  breaks = "default",
  title = "PM2.5",
  direction = 1,
  timezone = "UTC",
  verbose = TRUE
)
}
\arguments{
\item{raster}{A Raster\* object or a list of Raster\* objects.}

\item{index}{The index of the Raster* object. See details.}

\item{palette}{The color palette used to map cell values. This must be one
of the palettes available through \code{ggplot2::scale_colour_brewer()}.}

\item{breaks}{The breaks used to define the map cell color or \code{'default'}.}

\item{title}{(Optional) A plot title.}

\item{direction}{Numeric. \code{direction = -1} reverses color palette.}

\item{timezone}{The timezone to parse the dates. Default is UTC.}

\item{verbose}{Logical to display messages.}
}
\value{
a ggplot object
}
\description{
Creates a \pkg{ggplot2} plot object from a \code{Raster\*}
object. The returned plot object can be plotted or can be enhanced with
additional \pkg{ggplot2} intructions.
}
\details{
If a list of Raster\* objects is provided, a small-multiples plot
is created.

The \code{index} is typically associated with a time-axis or RasterLayer,
e.g \code{index = 1} is the first hour of a model.
}
\section{Methods (by class)}{
\itemize{
\item \code{Raster}: Method for Raster* objects.

\item \code{list}: Multi-threaded method for a list of Raster* objects.
}}

\examples{
\dontrun{
library(AirFireModeling)
setModelDataDir('~/Data/BlueSky')

# Load from server
rasterList <- raster_load(
  model = "PNW-4km",
  modelRun = c(2019100800, 2019100900, 2019101000, 2019101100),
  xlim = c(-125, -115),
  ylim = c(42, 50)
)

raster_map(rasterList, title="PNW-4km", index = 3)
}
}
