% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_subset.R
\name{raster_subset}
\alias{raster_subset}
\title{Subset a Raster* by distance from a location}
\usage{
raster_subset(raster, longitude, latitude, radius = 20000)
}
\arguments{
\item{raster}{A Raster* object.}

\item{longitude}{A target longitude.}

\item{latitude}{A target latitude.}

\item{radius}{A radius to subset by (in meters)}
}
\value{
a Raster* object
}
\description{
Subset a Raster* by distance from a location
}
