% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_subsetByPolygon.R
\name{raster_subsetByPolygon}
\alias{raster_subsetByPolygon}
\title{Subset a Raster* by polygon}
\usage{
raster_subsetByPolygon(raster = NULL, polygon = NULL)
}
\arguments{
\item{raster}{Raster\* object.}

\item{polygon}{Spatial\* object.}
}
\value{
A Raster* object.
}
\description{
This function subsets a raster object using cookie cutter
defined by \code{polygon}. All cells within the polygon are returned.
}
\examples{
\dontrun{
library(AirFireModeling)
setModelDataDir('~/Data/BlueSky')

# Load data
rasterList <- raster_load(
  model = "PNW-4km",
  modelRun = c(2019100800, 2019100900, 2019101000, 2019101100),
  xlim = c(-125, -116.5),
  ylim = c(42, 46.2)
)

library(MazamaSpatialUtils)
setSpatialDataDir("~/Data/Spatial")
loadSpatialData("USCensusStates")

OregonList <- raster_subsetByPolygon(
  rasterList,
  polygon = subset(USCensusStates, stateCode == "OR")
)

raster_map(
  OregonList,
  palette = "Spectral",
  breaks = c(-1, 0, 1, 2, 4, 8, 16, 32, 500),
  direction = -1
)
}
}
