% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{roc}
\alias{roc}
\title{ROC and AUC}
\usage{
roc(predicted, observed, showPlot = TRUE)
}
\arguments{
\item{predicted}{Predicted logical.}

\item{observed}{Observed logical.}

\item{showPlot}{Show plot logical.}
}
\value{
a AUROC list
}
\description{
ROC and AUC
}
